+++ lib/amazing_print/formatters/hash_formatter.rb
@@ -135,14 +135,23 @@ module AmazingPrint
 
         # Move the colon to the right side of the symbol
         key_string = key.inspect.include?('"') ? key.inspect.sub(':', '') : key.to_s
-        awesome_key = inspector.awesome(key).sub(/#{Regexp.escape(key.inspect)}/, "#{key_string}:")
+        awesome_key = format_key(key).sub(/#{Regexp.escape(key.inspect)}/, "#{key_string}:")
 
         "#{align(awesome_key, width)} #{inspector.awesome(value)}"
       end
 
       def pre_ruby19_syntax(key, value, width)
-        awesome_key = single_line { inspector.awesome(key) }
+        awesome_key = single_line { format_key(key) }
+
         "#{align(awesome_key, width)}#{colorize(' => ', :hash)}#{inspector.awesome(value)}"
+      end
+
+      def format_key(key)
+        if options[:format_keys]
+          inspector.awesome(key)
+        else
+          key.inspect
+        end
       end
 
       def plain_single_line(&)